Level1/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0